Fix tests affected by moving 'tests' folder on "[Migration][DO NOT MERGE] Fix and enable all new tests"
- Somewhat unrelated: moved "tests" folder out of "onnxscript".
- Updated import paths in copied code.
- Updated testdata lookup path due to relative path changes.
- Discovered an issue with installing via "pip install -e .". #1327
- Fixed a few issues of code under "ir" that doesn't work under python 3.8, or onnx w/ ir_version < 10.

All CI tests except for linter are expected to pass.

[ghstack-poisoned]
