+++ onnxscript/converter_test.py
@@ -29,7 +29,7 @@ import onnxscript.testing
 from onnxscript import BOOL, FLOAT, INT64, converter, graph, script, tensor
 from onnxscript.onnx_opset import opset11 as op11
 from onnxscript.onnx_opset import opset15 as op
-from onnxscript.tests.common import onnx_script_test_case, testutils
+from tests.common import onnx_script_test_case, testutils
 
 TEST_INPUT_DIR = pathlib.Path(__file__).parent / "tests" / "models"
 TEST_OUTPUT_DIR = TEST_INPUT_DIR / "testoutputs"
@@ -132,7 +132,7 @@ class TestConverter(testutils.TestBase):
                     self.check_run(val.function, val.input, val.output[0])
 
     def test_eager_op(self):
-        from onnxscript.tests.models import eager_op
+        from tests.models import eager_op
 
         test_functions = self.validate_save(eager_op, check_ort=True)
 
@@ -195,39 +195,39 @@ class TestConverter(testutils.TestBase):
         self.assertEqual(output_value_info.type.tensor_type.elem_type, onnx.TensorProto.FLOAT)
 
     def test_onnxfns1(self):
-        from onnxscript.tests.models import onnxfns1
+        from tests.models import onnxfns1
 
         self.validate(onnxfns1)
 
     def test_onnxfns1A(self):
-        from onnxscript.tests.models import onnxfns1A
+        from tests.models import onnxfns1A
 
         self.validate(onnxfns1A)
 
     def test_ort_custom_ops(self):
-        from onnxscript.tests.functions import ort_custom_ops
+        from tests.functions import ort_custom_ops
 
         self.validate(ort_custom_ops)
 
     def test_unary_op(self):
-        from onnxscript.tests.models import m1
+        from tests.models import m1
 
         self.validate_save(m1)
 
     def test_subfunction_check_model(self):
-        from onnxscript.tests.models import subfunction
+        from tests.models import subfunction
 
         model = subfunction.MyElu.function_ir.to_model_proto(producer_name="p2o")
         model = onnx.shape_inference.infer_shapes(model)
         onnx.checker.check_model(model)
 
     def test_subfunction(self):
-        from onnxscript.tests.models import subfunction
+        from tests.models import subfunction
 
         self.validate_save(subfunction, check_ort=True)
 
     def test_if_models(self):
-        from onnxscript.tests.models import if_statement
+        from tests.models import if_statement
 
         self.validate_save(if_statement)
 
@@ -246,28 +246,28 @@ class TestConverter(testutils.TestBase):
         self.assertEqual(proto.doc_string.strip(), "Combines ReduceSum, ReduceProd.")
 
     def test_signal(self):
-        from onnxscript.tests.models import signal_dft
+        from tests.models import signal_dft
 
         # shape_inference crashes on stft.
         self.validate_save(signal_dft, shape_inference=False)
 
     def test_multi(self):
-        from onnxscript.tests.models import multi
+        from tests.models import multi
 
         self.validate_save(multi, shape_inference=False)
 
     def test_dropout(self):
-        from onnxscript.tests.models import dropout
+        from tests.models import dropout
 
         self.validate_save(dropout, shape_inference=False)
 
     def test_attrref(self):
-        from onnxscript.tests.models import attrref
+        from tests.models import attrref
 
         self.validate_save(attrref, shape_inference=False)
 
     def test_renaming(self):
-        from onnxscript.tests.models import renaming
+        from tests.models import renaming
 
         self.validate_save(renaming, shape_inference=False)
 
@@ -276,18 +276,18 @@ class TestConverter(testutils.TestBase):
         reason="default_opset must be specified in script for functions that do not contain any use of an ONNX op",
     )
     def test_opt_output(self):
-        from onnxscript.tests.models import opt_output
+        from tests.models import opt_output
 
         self.validate_save(opt_output, shape_inference=False)
 
     def test_opt_input(self):
-        from onnxscript.tests.models import opt_input
+        from tests.models import opt_input
 
         self.validate_save(opt_input, shape_inference=False)
 
     @unittest.skip("A function with attributes cannot be exported as a model.")
     def test_onnxfns2(self):
-        from onnxscript.tests.models import onnxfns2
+        from tests.models import onnxfns2
 
         self.validate_save(onnxfns2, shape_inference=False)
 
@@ -301,7 +301,7 @@ class TestConverter(testutils.TestBase):
         self.validate_save(clipmax)
 
     def test_type_double(self):
-        from onnxscript.tests.models import type_double
+        from tests.models import type_double
 
         fcts = self.validate_save(type_double, check_ort=False)
         f = fcts["double_abs"]
@@ -320,17 +320,17 @@ class TestConverter(testutils.TestBase):
         self.validate_save(type_double, check_ort=True)
 
     def test_cast_like(self):
-        from onnxscript.tests.models import cast_like
+        from tests.models import cast_like
 
         self.validate_expansion(cast_like)
 
     def test_identity(self):
-        from onnxscript.tests.models import identity
+        from tests.models import identity
 
         self.validate_expansion(identity)
 
     def test_opset_import(self):
-        from onnxscript.tests.models import different_opset
+        from tests.models import different_opset
 
         fcts = self.validate_save(different_opset, shape_inference=False)
         s16 = str(fcts["shape_A"])
@@ -345,7 +345,7 @@ class TestConverter(testutils.TestBase):
         self.assertNotIn("version: 15", sdef)
 
     def test_sequences(self):
-        from onnxscript.tests.models import sequences
+        from tests.models import sequences
 
         test_functions = self.validate_save(sequences, check_ort=True)
 
@@ -372,7 +372,7 @@ class TestConverter(testutils.TestBase):
         np.testing.assert_almost_equal(eager_mode, result)
 
     def test_loops_break(self):
-        from onnxscript.tests.models import loops_break
+        from tests.models import loops_break
 
         test_functions = self.validate_save(loops_break, check_ort=True)
         self.assertIn("loop1", test_functions)
@@ -392,7 +392,7 @@ class TestConverter(testutils.TestBase):
         self.assertEqual(y.tolist(), [0, 11, -22])
 
     def test_loops_while(self):
-        from onnxscript.tests.models import loops_while
+        from tests.models import loops_while
 
         test_functions = self.validate_save(loops_while, check_ort=True)
         self.assertIn("loop1", test_functions)
@@ -409,7 +409,7 @@ class TestConverter(testutils.TestBase):
         self.assertEqual(res.tolist(), [0, 10, -20])
 
     def test_getitem(self):
-        from onnxscript.tests.models import getitem
+        from tests.models import getitem
 
         self.validate_save(getitem, check_ort=True, skip_check_ort=None)
         self.validate_run(getitem)
@@ -459,28 +459,28 @@ class TestConverter(testutils.TestBase):
         np.testing.assert_equal(output, expected_output)
 
     def test_graph_attr_scan(self):
-        from onnxscript.tests.models.graph_attr import cumulative_sum
+        from tests.models.graph_attr import cumulative_sum
 
         inputs = [np.array([1, 2, 3, 4, 5], dtype=np.int64)]
         expected_output = np.array([1, 3, 6, 10, 15], dtype=np.int64)
         self.check_run(cumulative_sum, inputs, expected_output)
 
     def test_graph_attr_loop(self):
-        from onnxscript.tests.models.graph_attr import sum_to
+        from tests.models.graph_attr import sum_to
 
         inputs = [np.array(6, dtype=np.int64)]
         expected_output = np.array([0, 1, 3, 6, 10, 15], dtype=np.int64)
         self.check_run(sum_to, inputs, expected_output)
 
     def test_graph_attr_loop_error(self):
-        from onnxscript.tests.models.graph_attr import sum_to_error
+        from tests.models.graph_attr import sum_to_error
 
         input = np.array(6, dtype=np.int64)
         with self.assertRaisesRegex(TypeError, "@graph"):
             sum_to_error(input)
 
     def test_loop_outer_scope(self):
-        from onnxscript.tests.models.graph_attr import loop_add
+        from tests.models.graph_attr import loop_add
 
         input_x = np.array([1, 2, 3], dtype=np.int64)
         input_m = np.array(3, dtype=np.int64)
@@ -504,7 +504,7 @@ class TestConverter(testutils.TestBase):
                 return op.DummyOp(body=inner)
 
     def test_attr(self):
-        from onnxscript.tests.functions import attr_test
+        from tests.functions import attr_test
 
         self.validate_run(attr_test)
