+++ onnxscript/type_annotation_test.py
@@ -12,7 +12,7 @@ import onnxscript
 import onnxscript.testing
 from onnxscript import FLOAT, INT64, script, type_annotation
 from onnxscript.onnx_opset import opset15 as op
-from onnxscript.tests.common import testutils
+from tests.common import testutils
 
 
 class TypeAnnotationTest(testutils.TestBase):
