+++ tests/functions/onnxfns2_test.py
@@ -1,7 +1,7 @@
 import unittest
 
-from onnxscript.tests.common import onnx_script_test_case
-from onnxscript.tests.models import onnxfns2
+from tests.common import onnx_script_test_case
+from tests.models import onnxfns2
 
 
 class TestOnnxFns(onnx_script_test_case.OnnxScriptTestCase):
